Add newly generated libGDX files as the base for the game.
This project was generated to work with IntelliJ, other IDE could be used, but the project has to be exported which can cause merge issues. The existing java file has not been modified. Additional java classes of any kind should be placed in ~\w19Wozniak\core\src\ca\mymacewan\memorygame